form of create exclude prevenDefault

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminRestController.java
@@ -54,7 +54,13 @@ public class AdminRestController {
 
     @DeleteMapping("users/{id}")
     public ResponseEntity<List<User>> deleteUser(@PathVariable("id") long id) {
-            userService.removeUser(id);
+          if(userService.getUserById(id)!=null){
+              userService.removeUser(id);
+
+          } else {
+              throw new RuntimeException("User not found");
+          }
+
         return ResponseEntity.ok(userService.getAllUsers());
     }
 }
